HexGrid
I split the hexGrid file from sprint 1 into two, HexCube, which I uploaded before this and this file. I changed the colour in backend for processing whenever a move was made and I made validate move which validates a move by checking if there are any neighbours of the same colour that the player is.